backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy>=2.0.36,<3.0.0
psycopg2-binary>=2.9.10,<3.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
email-validator>=2.2.0,<3.0.0
python-multipart>=0.0.12,<1.0.0
httpx>=0.27.0,<1.0.0
numpy>=2.0.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
passlib>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
